Progress Update!
ok! so i dont even know what i did but the tierLength can print to the console from score() and using some devious trickery we can run the score() function and save it to the "points" value on the website.

the issue now is that this "points" value is only updated when the site first loads, and the fix for this would be to update it every time the "selected" variable in the site changes (aka when we click on a different level).

this all is a VERYYYYYYY complicated thing but it's only complicated because of how these functions work, i blame javascript. although this """"""works"""""", if anyone has an idea on how to do this more simply then lmk lol.